+++ citrixadc/provider.go
@@ -187,6 +187,9 @@ func providerResources() map[string]*schema.Resource {
 		"citrixadc_appfwprofile_crosssitescripting_binding": resourceCitrixAdcAppfwprofile_crosssitescripting_binding(),
 		"citrixadc_appfwprofile_sqlinjection_binding":       resourceCitrixAdcAppfwprofile_sqlinjection_binding(),
 		"citrixadc_lbvserver_servicegroup_binding":          resourceCitrixAdcLbvserver_servicegroup_binding(),
+		"citrixadc_lbparameter":                             resourceCitrixAdcLbparameter(),
+		"citrixadc_iptunnel":                                resourceCitrixAdcIptunnel(),
+		"citrixadc_vlan":                                    resourceCitrixAdcVlan(),
 	}
 }